64: move code to app folder, add missing singletons
