+++ app/components/conditional_icon.php
