+++ app/components/finish_icon.php
