+++ app/pages/events/event_delete.php
