+++ app/pages/users/family_view.php
@@ -29,9 +29,9 @@ $add_member_list = em()->createQueryBuilder()
     ->orderBy('u.last_name')
     ->getQuery()->getArrayResult();
 
-$profile_picture = (file_exists("images/profile/" . $user->id)) ?
-    "/images/profile/" . $user->id
-    : "/images/profile/none.jpg";
+$profile_picture = "/app/" . (file_exists("app/images/profile/" . $user->id) ?
+    "images/profile/" . $user->id
+    : "images/profile/none.jpg");
 
 page($family->name)->css("family_list.css") ?>
 <?php if (check_auth(Access::$EDIT_USERS)): ?>
