+++ app/pages/users/take_user_control.php
